[refactor] working... somehow
