+++ fibsem/ui/qtdesigner_files/FibsemMillingWidget.py
@@ -17,6 +17,9 @@ class Ui_Form(object):
         Form.resize(597, 757)
         self.gridLayout = QtWidgets.QGridLayout(Form)
         self.gridLayout.setObjectName("gridLayout")
+        self.pushButton_stop_milling = QtWidgets.QPushButton(Form)
+        self.pushButton_stop_milling.setObjectName("pushButton_stop_milling")
+        self.gridLayout.addWidget(self.pushButton_stop_milling, 26, 1, 1, 1)
         self.frame = QtWidgets.QFrame(Form)
         self.frame.setMinimumSize(QtCore.QSize(0, 0))
         self.frame.setObjectName("frame")
@@ -24,18 +27,18 @@ class Ui_Form(object):
         self.gridLayout_4.setObjectName("gridLayout_4")
         self.label_milling_instructions = QtWidgets.QLabel(self.frame)
         self.label_milling_instructions.setObjectName("label_milling_instructions")
-        self.gridLayout_4.addWidget(self.label_milling_instructions, 24, 0, 1, 2)
+        self.gridLayout_4.addWidget(self.label_milling_instructions, 25, 0, 1, 2)
         self.pushButton_add_milling_stage = QtWidgets.QPushButton(self.frame)
         self.pushButton_add_milling_stage.setObjectName("pushButton_add_milling_stage")
         self.gridLayout_4.addWidget(self.pushButton_add_milling_stage, 3, 0, 1, 1)
         self.label_milling_stage = QtWidgets.QLabel(self.frame)
         self.label_milling_stage.setObjectName("label_milling_stage")
         self.gridLayout_4.addWidget(self.label_milling_stage, 2, 0, 1, 1)
-        self.scrollArea = QtWidgets.QScrollArea(self.frame)
-        self.scrollArea.setWidgetResizable(True)
-        self.scrollArea.setObjectName("scrollArea")
+        self.scrollArea_milling_stage = QtWidgets.QScrollArea(self.frame)
+        self.scrollArea_milling_stage.setWidgetResizable(True)
+        self.scrollArea_milling_stage.setObjectName("scrollArea_milling_stage")
         self.scrollAreaWidgetContents = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents.setGeometry(QtCore.QRect(0, 0, 545, 663))
+        self.scrollAreaWidgetContents.setGeometry(QtCore.QRect(0, -217, 545, 663))
         self.scrollAreaWidgetContents.setObjectName("scrollAreaWidgetContents")
         self.gridLayout_2 = QtWidgets.QGridLayout(self.scrollAreaWidgetContents)
         self.gridLayout_2.setObjectName("gridLayout_2")
@@ -127,35 +130,35 @@ class Ui_Form(object):
         self.doubleSpinBox_milling_current.setObjectName("doubleSpinBox_milling_current")
         self.gridLayout_3.addWidget(self.doubleSpinBox_milling_current, 1, 1, 1, 1)
         self.gridLayout_2.addWidget(self.groupBox_milling_settings, 0, 0, 1, 1)
-        self.groupBox_patterns = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
-        self.groupBox_patterns.setObjectName("groupBox_patterns")
-        self.gridLayout_5 = QtWidgets.QGridLayout(self.groupBox_patterns)
+        self.groupBox_pattern_settings = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
+        self.groupBox_pattern_settings.setObjectName("groupBox_pattern_settings")
+        self.gridLayout_5 = QtWidgets.QGridLayout(self.groupBox_pattern_settings)
         self.gridLayout_5.setObjectName("gridLayout_5")
-        self.doubleSpinBox_centre_x = QtWidgets.QDoubleSpinBox(self.groupBox_patterns)
+        self.doubleSpinBox_centre_x = QtWidgets.QDoubleSpinBox(self.groupBox_pattern_settings)
         self.doubleSpinBox_centre_x.setMinimum(-1e+28)
         self.doubleSpinBox_centre_x.setMaximum(1e+18)
         self.doubleSpinBox_centre_x.setSingleStep(0.1)
         self.doubleSpinBox_centre_x.setObjectName("doubleSpinBox_centre_x")
         self.gridLayout_5.addWidget(self.doubleSpinBox_centre_x, 1, 1, 1, 1)
-        self.label_pattern_set = QtWidgets.QLabel(self.groupBox_patterns)
+        self.label_pattern_set = QtWidgets.QLabel(self.groupBox_pattern_settings)
         self.label_pattern_set.setObjectName("label_pattern_set")
         self.gridLayout_5.addWidget(self.label_pattern_set, 0, 0, 1, 1)
-        self.doubleSpinBox_centre_y = QtWidgets.QDoubleSpinBox(self.groupBox_patterns)
+        self.doubleSpinBox_centre_y = QtWidgets.QDoubleSpinBox(self.groupBox_pattern_settings)
         self.doubleSpinBox_centre_y.setMinimum(-1e+16)
         self.doubleSpinBox_centre_y.setMaximum(1e+23)
         self.doubleSpinBox_centre_y.setSingleStep(0.1)
         self.doubleSpinBox_centre_y.setObjectName("doubleSpinBox_centre_y")
         self.gridLayout_5.addWidget(self.doubleSpinBox_centre_y, 1, 2, 1, 1)
-        self.label_centre_x = QtWidgets.QLabel(self.groupBox_patterns)
+        self.label_centre_x = QtWidgets.QLabel(self.groupBox_pattern_settings)
         self.label_centre_x.setObjectName("label_centre_x")
         self.gridLayout_5.addWidget(self.label_centre_x, 1, 0, 1, 1)
-        self.comboBox_patterns = QtWidgets.QComboBox(self.groupBox_patterns)
+        self.comboBox_patterns = QtWidgets.QComboBox(self.groupBox_pattern_settings)
         self.comboBox_patterns.setObjectName("comboBox_patterns")
         self.gridLayout_5.addWidget(self.comboBox_patterns, 0, 1, 1, 2)
         self.gridLayout_patterns = QtWidgets.QGridLayout()
         self.gridLayout_patterns.setObjectName("gridLayout_patterns")
         self.gridLayout_5.addLayout(self.gridLayout_patterns, 2, 0, 1, 3)
-        self.gridLayout_2.addWidget(self.groupBox_patterns, 2, 0, 1, 1)
+        self.gridLayout_2.addWidget(self.groupBox_pattern_settings, 2, 0, 1, 1)
         self.groupBox_options = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
         self.groupBox_options.setObjectName("groupBox_options")
         self.gridLayout_6 = QtWidgets.QGridLayout(self.groupBox_options)
@@ -180,8 +183,8 @@ class Ui_Form(object):
         self.checkBox_drift_correction_enabled.setObjectName("checkBox_drift_correction_enabled")
         self.gridLayout_7.addWidget(self.checkBox_drift_correction_enabled, 0, 0, 1, 1)
         self.gridLayout_2.addWidget(self.groupBox_drift_correction, 1, 0, 1, 1)
-        self.scrollArea.setWidget(self.scrollAreaWidgetContents)
-        self.gridLayout_4.addWidget(self.scrollArea, 4, 0, 4, 2)
+        self.scrollArea_milling_stage.setWidget(self.scrollAreaWidgetContents)
+        self.gridLayout_4.addWidget(self.scrollArea_milling_stage, 4, 0, 4, 2)
         self.pushButton_remove_milling_stage = QtWidgets.QPushButton(self.frame)
         self.pushButton_remove_milling_stage.setObjectName("pushButton_remove_milling_stage")
         self.gridLayout_4.addWidget(self.pushButton_remove_milling_stage, 3, 1, 1, 1)
@@ -189,16 +192,18 @@ class Ui_Form(object):
         self.comboBox_milling_stage.setObjectName("comboBox_milling_stage")
         self.gridLayout_4.addWidget(self.comboBox_milling_stage, 2, 1, 1, 1)
         self.gridLayout.addWidget(self.frame, 16, 0, 1, 2)
-        self.pushButton_stop_milling = QtWidgets.QPushButton(Form)
-        self.pushButton_stop_milling.setObjectName("pushButton_stop_milling")
-        self.gridLayout.addWidget(self.pushButton_stop_milling, 26, 0, 1, 2)
         self.progressBar_milling = QtWidgets.QProgressBar(Form)
         self.progressBar_milling.setProperty("value", 24)
         self.progressBar_milling.setObjectName("progressBar_milling")
-        self.gridLayout.addWidget(self.progressBar_milling, 27, 0, 1, 1)
+        self.gridLayout.addWidget(self.progressBar_milling, 27, 0, 1, 2)
+        self.pushButton_pause_milling = QtWidgets.QPushButton(Form)
+        self.pushButton_pause_milling.setObjectName("pushButton_pause_milling")
+        self.gridLayout.addWidget(self.pushButton_pause_milling, 26, 0, 1, 1)
         self.pushButton_run_milling = QtWidgets.QPushButton(Form)
         self.pushButton_run_milling.setObjectName("pushButton_run_milling")
         self.gridLayout.addWidget(self.pushButton_run_milling, 25, 0, 1, 2)
+        spacerItem = QtWidgets.QSpacerItem(20, 40, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
+        self.gridLayout.addItem(spacerItem, 28, 0, 1, 2)
 
         self.retranslateUi(Form)
         QtCore.QMetaObject.connectSlotsByName(Form)
@@ -210,6 +215,7 @@ class Ui_Form(object):
     def retranslateUi(self, Form):
         _translate = QtCore.QCoreApplication.translate
         Form.setWindowTitle(_translate("Form", "Form"))
+        self.pushButton_stop_milling.setText(_translate("Form", "Stop Milling"))
         self.label_milling_instructions.setText(_translate("Form", "Controls:"))
         self.pushButton_add_milling_stage.setText(_translate("Form", "Add"))
         self.label_milling_stage.setText(_translate("Form", "Milling Stage"))
@@ -224,7 +230,7 @@ class Ui_Form(object):
         self.label_spot_size.setText(_translate("Form", "Spot Size (um)"))
         self.label_dwell_time.setText(_translate("Form", "Dwell Time (us)"))
         self.label_spacing.setText(_translate("Form", "Spacing"))
-        self.groupBox_patterns.setTitle(_translate("Form", "Pattern"))
+        self.groupBox_pattern_settings.setTitle(_translate("Form", "Pattern"))
         self.label_pattern_set.setText(_translate("Form", "Pattern"))
         self.label_centre_x.setText(_translate("Form", "Centre (um)"))
         self.groupBox_options.setTitle(_translate("Form", "Options"))
@@ -234,5 +240,5 @@ class Ui_Form(object):
         self.groupBox_drift_correction.setTitle(_translate("Form", "Drift Correction"))
         self.checkBox_drift_correction_enabled.setText(_translate("Form", "Enabled"))
         self.pushButton_remove_milling_stage.setText(_translate("Form", "Remove"))
-        self.pushButton_stop_milling.setText(_translate("Form", "Stop Milling"))
+        self.pushButton_pause_milling.setText(_translate("Form", "Pause Milling"))
         self.pushButton_run_milling.setText(_translate("Form", "Run Milling"))
