+++ fibsem/ui/qtdesigner_files/FibsemMillingWidget.ui
@@ -14,6 +14,13 @@
    <string>Form</string>
   </property>
   <layout class="QGridLayout" name="gridLayout">
+   <item row="26" column="1">
+    <widget class="QPushButton" name="pushButton_stop_milling">
+     <property name="text">
+      <string>Stop Milling</string>
+     </property>
+    </widget>
+   </item>
    <item row="16" column="0" colspan="2">
     <widget class="QFrame" name="frame">
      <property name="minimumSize">
@@ -23,7 +30,7 @@
       </size>
      </property>
      <layout class="QGridLayout" name="gridLayout_4">
-      <item row="24" column="0" colspan="2">
+      <item row="25" column="0" colspan="2">
        <widget class="QLabel" name="label_milling_instructions">
         <property name="text">
          <string>Controls:</string>
@@ -45,7 +52,7 @@
        </widget>
       </item>
       <item row="4" column="0" rowspan="4" colspan="2">
-       <widget class="QScrollArea" name="scrollArea">
+       <widget class="QScrollArea" name="scrollArea_milling_stage">
         <property name="widgetResizable">
          <bool>true</bool>
         </property>
@@ -244,7 +251,7 @@
            </widget>
           </item>
           <item row="2" column="0">
-           <widget class="QGroupBox" name="groupBox_patterns">
+           <widget class="QGroupBox" name="groupBox_pattern_settings">
             <property name="title">
              <string>Pattern</string>
             </property>
@@ -367,17 +374,17 @@
      </layout>
     </widget>
    </item>
-   <item row="26" column="0" colspan="2">
-    <widget class="QPushButton" name="pushButton_stop_milling">
-     <property name="text">
-      <string>Stop Milling</string>
-     </property>
-    </widget>
-   </item>
-   <item row="27" column="0">
+   <item row="27" column="0" colspan="2">
     <widget class="QProgressBar" name="progressBar_milling">
      <property name="value">
       <number>24</number>
+     </property>
+    </widget>
+   </item>
+   <item row="26" column="0">
+    <widget class="QPushButton" name="pushButton_pause_milling">
+     <property name="text">
+      <string>Pause Milling</string>
      </property>
     </widget>
    </item>
@@ -387,6 +394,19 @@
       <string>Run Milling</string>
      </property>
     </widget>
+   </item>
+   <item row="28" column="0" colspan="2">
+    <spacer name="verticalSpacer">
+     <property name="orientation">
+      <enum>Qt::Vertical</enum>
+     </property>
+     <property name="sizeHint" stdset="0">
+      <size>
+       <width>20</width>
+       <height>40</height>
+      </size>
+     </property>
+    </spacer>
    </item>
   </layout>
  </widget>
