+++ fibsem/ui/utils.py
@@ -1,162 +1,29 @@
 import logging
+import time
 from dataclasses import dataclass
-from typing import Union
+from pathlib import Path
+from typing import List, Tuple, Union
 
-import matplotlib.patches as mpatches
 import matplotlib.pyplot as plt
-import numpy as np
-from PIL import Image
-from typing import List, Tuple
-
-from fibsem import constants, conversions
-from fibsem.structures import Point, FibsemImage, FibsemRectangle
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg
-from matplotlib.figure import Figure
-from matplotlib.patches import Rectangle
-from PyQt5.QtWidgets import QMessageBox, QSizePolicy, QVBoxLayout, QWidget
-from fibsem.structures import (FibsemPatternSettings, 
-                               FibsemRectangleSettings,  FibsemLineSettings, 
-                               FibsemCircleSettings, FibsemBitmapSettings)
-from fibsem.microscope import FibsemMicroscope
 import napari
-from fibsem.utils import load_yaml, save_yaml
-from dataclasses import dataclass
-
-@dataclass
-class DrawnPattern:
-    pattern: np.ndarray
-    position: Point
-    is_exclusion: bool
-
-
-# # TODO: clean up and refactor these (_WidgetPlot and _PlotCanvas)
-# class _WidgetPlot(QWidget):
-#     def __init__(self, *args, display_image, **kwargs):
-#         QWidget.__init__(self, *args, **kwargs)
-#         self.setLayout(QVBoxLayout())
-#         self.canvas = _PlotCanvas(self, image=display_image)
-#         self.layout().addWidget(self.canvas)
-
-
-# class _PlotCanvas(FigureCanvasQTAgg):
-#     def __init__(self, parent=None, image=None):
-#         self.fig = Figure()
-#         FigureCanvasQTAgg.__init__(self, self.fig)
-
-#         self.setParent(parent)
-#         FigureCanvasQTAgg.setSizePolicy(
-#             self, QSizePolicy.Expanding, QSizePolicy.Expanding
-#         )
-#         FigureCanvasQTAgg.updateGeometry(self)
-#         self.image = image
-#         self.plot()
-#         self.createConn()
-
-#         self.figureActive = False
-#         self.axesActive = None
-#         self.cursorGUI = "arrow"
-#         self.cursorChanged = False
-
-#     def plot(self):
-#         gs0 = self.fig.add_gridspec(1, 1)
-
-#         self.ax11 = self.fig.add_subplot(gs0[0], xticks=[], yticks=[], title="")
-
-#         if self.image.ndim == 3:
-#             self.ax11.imshow(self.image,)
-#         else:
-#             self.ax11.imshow(self.image, cmap="gray")
-
-#     def updateCanvas(self, event=None):
-#         ax11_xlim = self.ax11.get_xlim()
-#         ax11_xvis = ax11_xlim[1] - ax11_xlim[0]
-
-#         while len(self.ax11.patches) > 0:
-#             [p.remove() for p in self.ax11.patches]
-#         while len(self.ax11.texts) > 0:
-#             [t.remove() for t in self.ax11.texts]
-
-#         ax11_units = ax11_xvis * 0.003
-#         self.fig.canvas.draw()
-
-#     def createConn(self):
-#         self.fig.canvas.mpl_connect("figure_enter_event", self.activeFigure)
-#         self.fig.canvas.mpl_connect("figure_leave_event", self.leftFigure)
-#         self.fig.canvas.mpl_connect("button_press_event", self.mouseClicked)
-#         self.ax11.callbacks.connect("xlim_changed", self.updateCanvas)
-
-#     def activeFigure(self, event):
-#         self.figureActive = True
-
-#     def leftFigure(self, event):
-#         self.figureActive = False
-#         if self.cursorGUI != "arrow":
-#             self.cursorGUI = "arrow"
-#             self.cursorChanged = True
-
-#     def mouseClicked(self, event):
-#         x = event.xdata
-#         y = event.ydata
-
-
-# @dataclass
-# class Crosshair:
-#     rectangle_horizontal: plt.Rectangle
-#     rectangle_vertical: plt.Rectangle
-
-
-# # TODO update with Point
-# def draw_crosshair(image, canvas, x: float = None, y: float = None, colour: str ="yellow"):
-#     # draw crosshairs
-#     crosshair = create_crosshair(image, x, y, colour=colour)
-#     for patch in crosshair.__dataclass_fields__:
-#         canvas.ax11.add_patch(getattr(crosshair, patch))
-#         getattr(crosshair, patch).set_visible(True)
-
-# # draw arrow
-# def draw_arrow(p1: Point, p2: Point, canvas) -> None:
-#     """Draw an arrow patch from p1 to p2"""
-#     x1, y1 = p1.x, p1.y
-#     x2, y2 = p2.x, p2.y
-#     line = mpatches.Arrow(x1, y1, x2 - x1, y2 - y1, color="white")
-
-#     # draw line
-#     canvas.ax11.add_patch(line)
-#     return
-
-
-# def display_error_message(message, title="Error"):
-#     """PyQt dialog box displaying an error message."""
-#     logging.exception(message)
-
-#     error_dialog = QMessageBox()
-#     error_dialog.setIcon(QMessageBox.Critical)
-#     error_dialog.setText(message)
-#     error_dialog.setWindowTitle(title)
-#     error_dialog.exec_()
-
-#     return error_dialog
-
-
-# def message_box_ui(title: str, text: str, buttons = QMessageBox.Yes | QMessageBox.No):
-
-#     msg = QMessageBox()
-#     msg.setWindowTitle(title)
-#     msg.setText(text)
-#     msg.setStandardButtons(buttons)
-#     msg.exec_()
-
-#     response = True if (msg.clickedButton() == msg.button(QMessageBox.Yes)) or (msg.clickedButton() == msg.button(QMessageBox.Ok) ) else False
-
-#     return response
-
-
-from PyQt5 import QtWidgets
-from PyQt5.QtCore import Qt
+import numpy as np
+from napari.layers import Layer as NapariLayer
+from napari.layers.shapes.shapes import Shapes as NapariShapes
+from PIL import Image
+from PyQt5 import QtGui, QtWidgets
 from PyQt5.QtGui import QImage, QPixmap
-from PyQt5.QtWidgets import QGridLayout, QLabel
+from PyQt5.QtWidgets import (
+    QLabel,
+    QMessageBox,
+)
 
-COLOURS = ["yellow", "cyan", "magenta", "lime", "orange", "hotpink", "green", "blue", "red", "purple"]
+from fibsem import config as cfg
+from fibsem import constants
+from fibsem.microscope import FibsemMicroscope
+from fibsem.milling import FibsemMillingStage
+from fibsem.structures import (
+    Point,
+)
 
 
 def set_arr_as_qlabel(
@@ -175,490 +42,6 @@ def set_arr_as_qlabel(
     return label
 
 
-# def set_arr_as_qlabel_8(
-#     arr: np.ndarray,
-#     label: QLabel,
-#     shape: tuple = (1536//4, 1024//4),
-# ) -> QLabel:
-
-#     image = QImage(
-#         arr.data,
-#         arr.shape[1],
-#         arr.shape[0],
-#         QImage.Format_Grayscale8,
-#     )
-#     label.setPixmap(QPixmap.fromImage(image).scaled(*shape))
-
-#     return label
-
-
-def convert_pattern_to_napari_circle(
-    pattern_settings: FibsemCircleSettings, image: FibsemImage
-):
-    
-    if not isinstance(pattern_settings, FibsemCircleSettings):
-        raise ValueError("Pattern is not a circle")
-    
-    # image centre
-    icy, icx = image.data.shape[0] // 2, image.data.shape[1] // 2
-    # pixel size
-    pixelsize_x, pixelsize_y = image.metadata.pixel_size.x, image.metadata.pixel_size.y
-
-
-    # pattern to pixel coords
-    r = int(pattern_settings.radius / pixelsize_x)
-    cx = int(icx + (pattern_settings.centre_x / pixelsize_y))
-    cy = int(icy - (pattern_settings.centre_y / pixelsize_y))
-
-    # create corner coords
-    xmin, ymin = cx - r, cy - r
-    xmax, ymax = cx + r, cy + r
-
-    # create circle
-    shape = [[ymin, xmin], [ymin, xmax], [ymax, xmax], [ymax, xmin]]  # ??
-    return shape
-
-
-def convert_pattern_to_napari_line(
-    pattern_settings: FibsemLineSettings, image: FibsemImage
-) -> np.ndarray:
-    # image centre
-    icy, icx = image.data.shape[0] // 2, image.data.shape[1] // 2
-    # pixel size
-    pixelsize_x, pixelsize_y = image.metadata.pixel_size.x, image.metadata.pixel_size.y
-    
-    # extract pattern information from settings
-    if not isinstance(pattern_settings, FibsemLineSettings):
-        raise ValueError("Pattern is not a line")
-
-    start_x = pattern_settings.start_x
-    start_y = pattern_settings.start_y
-    end_x = pattern_settings.end_x
-    end_y = pattern_settings.end_y
-
-    # pattern to pixel coords
-    px0 = int(icx + (start_x / pixelsize_x))
-    py0 = int(icy - (start_y / pixelsize_y))
-    px1 = int(icx + (end_x / pixelsize_x))
-    py1 = int(icy - (end_y / pixelsize_y))
-
-    # napari shape format [[y_start, x_start], [y_end, x_end]])
-    shape = [[py0, px0], [py1, px1]]
-    return shape
-
-def convert_pattern_to_napari_rect(
-    pattern_settings: FibsemRectangleSettings, image: FibsemImage
-) -> np.ndarray:
-    # image centre
-    icy, icx = image.data.shape[0] // 2, image.data.shape[1] // 2
-    # pixel size
-    pixelsize_x, pixelsize_y = image.metadata.pixel_size.x, image.metadata.pixel_size.y
-    # extract pattern information from settings
-    if isinstance(pattern_settings, FibsemLineSettings):
-        pattern_width = pattern_settings.end_x - pattern_settings.start_x
-        pattern_height = max(pattern_settings.end_y - pattern_settings.start_y, 0.5e-6)
-        pattern_rotation = np.arctan2(
-            pattern_height, pattern_width
-        )  # TODO: line rotation doesnt work correctly, fix
-        pattern_centre_x = (pattern_settings.end_x + pattern_settings.start_x) / 2
-        pattern_centre_y = (pattern_settings.end_y + pattern_settings.start_y) / 2
-
-    
-    elif isinstance(pattern_settings, FibsemCircleSettings): #only used for out of bounds check
-        pattern_width = 2*pattern_settings.radius
-        pattern_height = 2*pattern_settings.radius
-        pattern_centre_x = pattern_settings.centre_x
-        pattern_centre_y = pattern_settings.centre_y
-        pattern_rotation = 0
-
-    else:
-        pattern_width = pattern_settings.width
-        pattern_height = pattern_settings.height
-        pattern_centre_x = pattern_settings.centre_x
-        pattern_centre_y = pattern_settings.centre_y
-        pattern_rotation = pattern_settings.rotation
-    # pattern to pixel coords
-    w = int(pattern_width / pixelsize_x)
-    h = int(pattern_height / pixelsize_y)
-    cx = int(icx + (pattern_centre_x / pixelsize_y))
-    cy = int(icy - (pattern_centre_y / pixelsize_y))
-    r = -pattern_rotation  #
-    xmin, xmax = -w / 2, w / 2
-    ymin, ymax = -h / 2, h / 2
-    px0 = cx + (xmin * np.cos(r) - ymin * np.sin(r))
-    py0 = cy + (xmin * np.sin(r) + ymin * np.cos(r))
-    px1 = cx + (xmax * np.cos(r) - ymin * np.sin(r))
-    py1 = cy + (xmax * np.sin(r) + ymin * np.cos(r))
-    px2 = cx + (xmax * np.cos(r) - ymax * np.sin(r))
-    py2 = cy + (xmax * np.sin(r) + ymax * np.cos(r))
-    px3 = cx + (xmin * np.cos(r) - ymax * np.sin(r))
-    py3 = cy + (xmin * np.sin(r) + ymax * np.cos(r))
-    # napari shape format
-    shape = [[py0, px0], [py1, px1], [py2, px2], [py3, px3]]
-    return shape
-
-def create_crosshair_shape(centre_point: Point, image: FibsemImage,eb_image: FibsemImage):
-
-    icy, icx = image.data.shape[0] // 2, image.data.shape[1] // 2
-
-    pixelsize_x, pixelsize_y = image.metadata.pixel_size.x, image.metadata.pixel_size.y
-
-    pattern_centre_x = centre_point.x
-    pattern_centre_y = centre_point.y
-
-    cx = int(icx + (pattern_centre_x / pixelsize_y))
-    cy = int(icy - (pattern_centre_y / pixelsize_y))
-
-    r_angles = [0,np.deg2rad(90)] #
-    w = 40
-    h = 1
-    crosshair_shapes = []
-
-    for r in r_angles:
-        xmin, xmax = -w / 2, w / 2
-        ymin, ymax = -h / 2, h / 2
-        px0 = cx + (xmin * np.cos(r) - ymin * np.sin(r))
-        py0 = cy + (xmin * np.sin(r) + ymin * np.cos(r))
-        px1 = cx + (xmax * np.cos(r) - ymin * np.sin(r))
-        py1 = cy + (xmax * np.sin(r) + ymin * np.cos(r))
-        px2 = cx + (xmax * np.cos(r) - ymax * np.sin(r))
-        py2 = cy + (xmax * np.sin(r) + ymax * np.cos(r))
-        px3 = cx + (xmin * np.cos(r) - ymax * np.sin(r))
-        py3 = cy + (xmin * np.sin(r) + ymax * np.cos(r))
-        # napari shape format
-        shape = [[py0, px0], [py1, px1], [py2, px2], [py3, px3]]
-        if eb_image is not None:
-                for c in shape:
-                    c[1] += eb_image.data.shape[1]
-        crosshair_shapes.append(shape)
-
-    return crosshair_shapes
-
-
-
-
-
-
-def convert_bitmap_pattern_to_napari_image(
-        pattern_settings: FibsemBitmapSettings, image: FibsemImage
-) -> np.ndarray:
-    # image centre
-    icy, icx = image.data.shape[0] // 2, image.data.shape[1] // 2
-    # pixel size
-    pixelsize_x, pixelsize_y = image.metadata.pixel_size.x, image.metadata.pixel_size.y
-
-    resize_x = int(pattern_settings.width / pixelsize_x)
-    resize_y = int(pattern_settings.height / pixelsize_y)
-
-    
-    image_bmp = Image.open(pattern_settings.path)
-    image_resized = image_bmp.resize((resize_x, resize_y))
-    image_rotated = image_resized.rotate(-pattern_settings.rotation, expand=True)
-    img_array = np.array(image_rotated)
-
-    pattern_centre_x = int(icx - pattern_settings.width/pixelsize_x/2) + image.data.shape[1] 
-    pattern_centre_y = int(icy - pattern_settings.height/pixelsize_y/2)
-
-    pattern_point_x = int(pattern_centre_x + pattern_settings.centre_x / pixelsize_x)
-    pattern_point_y = int(pattern_centre_y - pattern_settings.centre_y / pixelsize_y)
-
-    translate_position = (pattern_point_y,pattern_point_x)
-
-    
-    return img_array, translate_position
-   
-def draw_annulus_shape(pattern_settings: FibsemCircleSettings, image: FibsemImage) -> DrawnPattern:
-    """Convert an annulus pattern to a np array. Note: annulus can only be plotted as image
-    Args:
-        pattern_settings: FibsemCircleSettings: Annulus pattern settings.
-        image: FibsemImage: Image to draw pattern on.
-    Returns:
-        np.ndarray: Annulus shape in image.
-        Point: Position of the annulus in the image.
-    """
-    
-    # image parameters (centre, pixel size)
-    icy, icx = image.data.shape[0] // 2, image.data.shape[1] // 2
-    pixelsize_x, pixelsize_y = image.metadata.pixel_size.x, image.metadata.pixel_size.y
-
-    # pattern parameters
-    radius = pattern_settings.radius
-    thickness = pattern_settings.thickness
-    center_x = pattern_settings.centre_x
-    center_y = pattern_settings.centre_y
-
-    radius_px = radius / pixelsize_x # isotropic
-    shape = int(2 * radius_px)
-    inner_radius_ratio = 0 # full circle
-    if not np.isclose(thickness, 0):
-        inner_radius_ratio = (radius - thickness)/radius
-   
-    annulus_shape = _create_annulus_shape(width=shape, height=shape, 
-                                          inner_radius=inner_radius_ratio, 
-                                          outer_radius=1)
-
-    # get pattern centre in image coordinates
-    pattern_centre_x = int(icx + center_x / pixelsize_x)
-    pattern_centre_y = int(icy - center_y / pixelsize_y)
-
-    pos = Point(x=pattern_centre_x, y=pattern_centre_y)
-
-    return DrawnPattern(pattern=annulus_shape, position=pos, is_exclusion=pattern_settings.is_exclusion)
-
-
-def draw_rectangle_shape(pattern_settings: FibsemRectangleSettings, image: FibsemImage) -> DrawnPattern:
-    """Convert a rectangle pattern to a np array.
-    Args:
-        pattern_settings: FibsemRectangleSettings: Rectangle pattern settings.
-        image: FibsemImage: Image to draw pattern on.
-    Returns:
-        np.ndarray: Rectangle shape in image.
-        Point: Position of the rectangle in the image.
-    """
-    # TODO: support rotation for drawing rectangles
-
-    # image parameters (centre, pixel size)
-    icy, icx = image.data.shape[0] // 2, image.data.shape[1] // 2
-    pixelsize_x, pixelsize_y = image.metadata.pixel_size.x, image.metadata.pixel_size.y
-
-    # pattern parameters
-    width = pattern_settings.width
-    height = pattern_settings.height
-    centre_x = pattern_settings.centre_x
-    centre_y = pattern_settings.centre_y
-    rotation = pattern_settings.rotation
-
-    # pattern to pixel coords
-    w = int(width / pixelsize_x)
-    h = int(height / pixelsize_y)
-    cx = int(icx + (centre_x / pixelsize_y))
-    cy = int(icy - (centre_y / pixelsize_y))
-
-    #
-    shape = np.ones((h, w))
-    shape[:h, :w] = 1
-
-    # get pattern centre in image coordinates
-    pos = Point(x=cx, y=cy)
-
-    return DrawnPattern(pattern=shape, position=pos, is_exclusion=pattern_settings.is_exclusion)
-
-def draw_pattern_shape(ps, image):
-    if isinstance(ps, FibsemCircleSettings):
-        return draw_annulus_shape(ps, image)
-    elif isinstance(ps, FibsemRectangleSettings):
-        return draw_rectangle_shape(ps, image)
-    else:
-        raise ValueError(f"Unsupported shape type {type(ps)}")
-
-def draw_pattern_in_image(image: np.ndarray, 
-                          drawn_pattern: DrawnPattern) -> np.ndarray:
-
-    pattern = drawn_pattern.pattern
-    pos = drawn_pattern.position
-    
-    # place the annulus shape in the image
-    w = pattern.shape[1] // 2
-    h = pattern.shape[0] // 2
-
-    # fill the annulus shape in the image
-    xmin, xmax = pos.x - w, pos.x + w
-    ymin, ymax = pos.y - h, pos.y + h
-    zero_image = np.zeros_like(image)
-    zero_image[ymin:ymax, xmin:xmax] = pattern[:2*h, :2*w].astype(bool)
-
-    # if the pattern is an exclusion, set the image to zero
-    if drawn_pattern.is_exclusion:
-        image[zero_image == 1] = 0
-    else:
-        # add the annulus shape to the image, clip to 1
-        image = np.clip(image+zero_image, 0, 1)
-
-    return image
-
-
-def compose_pattern_image(image: np.ndarray, drawn_patterns: List[DrawnPattern]) -> np.ndarray:
-    """Create an image with annulus shapes."""
-    # create an empty image
-    pattern_image = np.zeros_like(image)
-
-    # sort drawn_patterns so that exclusions are drawn last
-    drawn_patterns = sorted(drawn_patterns, key=lambda x: x.is_exclusion)
-
-    # add each pattern shape to the image
-    for dp in drawn_patterns:
-        pattern_image = draw_pattern_in_image(pattern_image, dp)
-
-    return pattern_image
-        
-
-def _create_annulus_shape(width, height, inner_radius, outer_radius):
-    # Create a grid of coordinates
-    x = np.linspace(-1, 1, width)
-    y = np.linspace(-1, 1, height)
-    X, Y = np.meshgrid(x, y)
-    distance = np.sqrt(X**2 + Y**2)
-    # Generate the donut shape
-    donut = np.logical_and(distance <= outer_radius, distance >= inner_radius).astype(int)
-    return donut
-
-IGNORE_SHAPES_LAYERS = ["ruler_line","crosshair","scalebar","scalebar_value", "label", "alignment_area"] # ignore these layers when removing all shapes
-IMAGE_PATTERN_LAYERS = ["annulus-layer", "bmp_Image"]
-
-def _remove_all_layers(viewer: napari.Viewer, layer_type = napari.layers.shapes.shapes.Shapes, _ignore: List[str] = []):
-
-    # remove all shapes layers
-    layers_to_remove = []
-    layers_to_ignore = IGNORE_SHAPES_LAYERS + _ignore
-    for layer in viewer.layers:
-
-        if layer.name in layers_to_ignore:
-            continue
-        if isinstance(layer, layer_type) or any([layer_name in layer.name for layer_name in IMAGE_PATTERN_LAYERS]):
-            layers_to_remove.append(layer)
-    for layer in layers_to_remove:
-        viewer.layers.remove(layer)  # Not removing the second layer?
-
-
-def _draw_patterns_in_napari(
-    viewer: napari.Viewer,
-    ib_image: FibsemImage,
-    eb_image: FibsemImage,
-    milling_stages: List["FibsemMillingStage"],
-    draw_crosshair: bool = True,
-):
-
-    # colour wheel
-    COLOURS = ["yellow", "cyan", "magenta", "lime", "orange", "hotpink", "green", "blue", "red", "purple"]
-
-    # convert fibsem patterns to napari shapes
-    import time
-
-    t_1 = time.time()
-    _ignore = []
-    for i, stage in enumerate(milling_stages):
-        shape_patterns = []
-        shape_types = []
-        t0 = time.time()
-
-        patterns: List[FibsemPatternSettings] = stage.pattern.define()
-        point: Point = stage.pattern.point
-        name: str = stage.name
-        is_line_pattern: bool = False
-        
-        # amnulus shapes
-        is_annulus: bool = False
-        annulus_layer = f"{name}-annulus-layer"
-
-        drawn_patterns = []
-
-        for pattern_settings in patterns:
-            # if isinstance(pattern_settings, FibsemBitmapSettings):
-            #     if pattern_settings.path == None or pattern_settings.path == '':
-            #         continue
-
-            #     bmp_Image, translate_position = convert_bitmap_pattern_to_napari_image(pattern_settings=pattern_settings, image=ib_image)
-            #     if "bmp_Image" in viewer.layers:
-            #         viewer.layers.remove(viewer.layers["bmp_Image"])
-            #     viewer.add_image(bmp_Image,translate=translate_position,name="bmp_Image")
-            #     shape_patterns = []
-            #     _ignore.append("bmp_Image")
-            #     continue
-
-            if isinstance(pattern_settings, FibsemCircleSettings):
-                if pattern_settings.thickness != 0:
-                    # # annulus is special case becaue napari can't draw annulus as shape, so we draw them as an image
-                    # is_annulus = True
-                    # drawn_patterns.append(draw_pattern_shape(pattern_settings, ib_image))
-                    continue
-                else:
-                    shape = convert_pattern_to_napari_circle(pattern_settings=pattern_settings, image=ib_image)
-                    shape_types.append("ellipse")
-                    _ignore.append(name)
-
-            elif isinstance(pattern_settings, FibsemLineSettings):
-                shape = convert_pattern_to_napari_line(pattern_settings=pattern_settings, image=ib_image)
-                shape_types.append("line")
-                _ignore.append(name)
-                is_line_pattern = True
-            
-            else:
-                shape = convert_pattern_to_napari_rect(
-                    pattern_settings=pattern_settings, image=ib_image
-                )
-                shape_types.append("rectangle")
-                _ignore.append(name)
-
-            # offset the x coord by image width
-            if eb_image is not None:
-                for c in shape:
-                    c[1] += eb_image.data.shape[1]
-            shape_patterns.append(shape)
-        
-        t1 = time.time()
-
-        if len(shape_patterns) > 0:
-            
-            if draw_crosshair:
-                crosshair_shapes = create_crosshair_shape(centre_point=point, image=ib_image, eb_image=eb_image)
-                crosshair_shape_types = ["rectangle","rectangle"]
-                shape_patterns += crosshair_shapes
-                shape_types += crosshair_shape_types
-
-            # _name = f"Stage {i+1:02d}"
-            if name in viewer.layers:
-                viewer.layers[name].data = []
-                viewer.layers[name].data = shape_patterns
-                viewer.layers[name].shape_type = shape_types
-                viewer.layers[name].edge_color = COLOURS[i % len(COLOURS)]
-                viewer.layers[name].face_color=COLOURS[i % len(COLOURS)]
-            else:
-                viewer.add_shapes(
-                    shape_patterns,
-                    name=name,
-                    shape_type=shape_types,
-                    edge_width=0.5,
-                    edge_color=COLOURS[i % len(COLOURS)],
-                    face_color=COLOURS[i % len(COLOURS)],
-                    opacity=0.5,
-                    blending="translucent",
-                )
-
-            if is_line_pattern:
-                viewer.layers[name].edge_width = 3
-
-        if is_annulus:
-            # draw all the annulus for the stage on the same image
-            if annulus_layer in viewer.layers:
-                viewer.layers.remove(viewer.layers[annulus_layer])
-
-            annulus_image = compose_pattern_image(ib_image.data, drawn_patterns)
-
-            translation = (0, eb_image.data.shape[1])
-            label_layer = viewer.add_labels(data=annulus_image, 
-                                translate=translation, 
-                                name=annulus_layer,
-                                blending="additive",
-                                opacity=0.6)
-            
-            cmap = {0: "black", 1: COLOURS[i % len(COLOURS)]}
-            if hasattr(label_layer, "colormap"): # attribute changed in napari 0.5.0+
-                label_layer.colormap = cmap
-            else:
-                label_layer.color = cmap
-            _ignore.append(annulus_layer)
-
-        t2 = time.time()
-        # remove all un-updated layers (assume they have been deleted)        
-        _remove_all_layers(viewer=viewer, layer_type=napari.layers.shapes.shapes.Shapes, _ignore=_ignore)
-        # TODO: annulus layers are not removed correctly
-        t3 = time.time()
-        logging.debug(f"_DRAW_SHAPES: CONVERT: {t1-t0}, ADD/UPDATE: {t2-t1}, REMOVE: {t3-t2}")
-    t_2 = time.time()
-    logging.debug(f"_DRAW_SHAPES: total time: {t_2-t_1}")
-
-    return _ignore # list of milling pattern layers
-
 def message_box_ui(title: str, text: str, buttons=QMessageBox.Yes | QMessageBox.No):
     msg = QMessageBox()
     msg.setWindowTitle(title)
@@ -675,7 +58,9 @@ def message_box_ui(title: str, text: str, buttons=QMessageBox.Yes | QMessageBox.
 
     return response
 
-from PyQt5 import QtGui
+
+
+
 def _display_logo(path, label, shape=[50, 50]):
     label.setScaledContents(True)
     label.setFixedSize(*shape)
@@ -846,42 +231,8 @@ def _draw_scalebar(viewer: napari.Viewer, eb_image, ib_image,is_checked=False, w
             viewer.layers["scalebar_value"].opacity = 0
 
 
-    
-
-
-def convert_point_to_napari(resolution: list, pixel_size: float, centre: Point):
-    icy, icx = resolution[1] // 2, resolution[0] // 2
-
-    cx = int(icx + (centre.x / pixel_size))
-    cy = int(icy - (centre.y / pixel_size))
-
-    return Point(cx, cy)
-
-
-def validate_pattern_placement(
-    resolution: Tuple[int, int], shape: List[List[float]]
-):
-    x_lim = resolution[0]
-    y_lim = resolution[1]
-
-    for coordinate in shape:
-        x_coord = coordinate[1]
-        y_coord = coordinate[0]
-
-        if x_coord < 0 or x_coord > x_lim:
-            return False
-        if y_coord < 0 or y_coord > y_lim:
-            return False
-
-    return True
-
-
-from fibsem import config as cfg
-from pathlib import Path
-
 
 # TODO: add filters for file types
-
 
 def _get_directory_ui(
     msg: str = "Select a directory", path: Path = cfg.LOG_PATH, parent=None
@@ -932,28 +283,6 @@ def _get_text_ui(
     )
 
     return text, okPressed
-
-def _draw_milling_stages_on_image(image: FibsemImage, milling_stages: List["FibsemMillingStage"], show: bool = True):
-
-    viewer = napari.Viewer()
-    viewer.add_image(image.data, name='test_image')
-    _draw_patterns_in_napari(viewer=viewer,ib_image=image,eb_image=None,milling_stages=milling_stages)
-    screenshot = viewer.screenshot()
-    fig, ax = plt.subplots(figsize=(10, 10))
-    ax.imshow(screenshot)
-    viewer.close()
-
-    for i,stage in enumerate(milling_stages):
-    
-        plt.plot(0,0,'-',color=COLOURS[i % len(COLOURS)],label=stage.name)
-
-    ax.axis('off')
-    ax.legend()
-    if show:
-        plt.show()
-    
-    return fig
-
 
 def show_information_dialog(microscope: FibsemMicroscope, parent=None):
     import fibsem
